Move SearchStartTask to SearchResultsFragment. Should fix #219
