+++ res/layout/activity_searchresult_list.xml
@@ -1,11 +1,6 @@
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/searchresult_list"
-    android:name="de.geeksfactory.opacclient.frontend.SearchResultListFragment"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/searchresult_list_container"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+   	android:layout_height="match_parent"
     android:layout_marginLeft="16dp"
-    android:layout_marginRight="16dp"
-    tools:context=".SearchResultListActivity"
-    tools:layout="@android:layout/list_content" />
-
+    android:layout_marginRight="16dp" />
